Use proper type casting in AM_ALIGNOF_PTR() and AM_ALIGN_SIZE() implementations

+++ libs/onesize/onesize.c
@@ -201,7 +201,7 @@ void am_onesize_ctor(struct am_onesize *hnd, const struct am_onesize_cfg *cfg) {
     hnd->pool = cfg->pool;
     hnd->block_size =
         AM_MAX(cfg->block_size, (int)sizeof(struct am_slist_item));
-    hnd->block_size = AM_ALIGN_SIZE(hnd->block_size, alignment);
+    hnd->block_size = (int)AM_ALIGN_SIZE(hnd->block_size, alignment);
     hnd->nbump = 0;
 
     AM_ASSERT(hnd->pool.size >= hnd->block_size);
